refactor(pool): remove all sessions-related methods out of pool
NODE-1892
